Abstract common Horovod operations from any specific framework (#185)
* Abstract common Horovod operations from any specific framework

* Bump version to 0.12.0dev0 due to higher-risk upgrade
