+++ horovod/common/common.cc
@@ -1,0 +1,98 @@
+// Copyright 2018 Uber Technologies, Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// =============================================================================
+
+#include <sstream>
+#include <assert.h>
+
+#include "common.h"
+
+namespace horovod {
+namespace common {
+
+Status::Status() {
+}
+
+Status::Status(StatusType type, std::string reason) {
+  type_ = type;
+  reason_ = reason;
+}
+
+Status Status::OK() {
+  return Status();
+}
+
+Status Status::UnknownError(std::string message) {
+  return Status(StatusType::UNKNOWN_ERROR, message);
+}
+
+Status Status::PreconditionError(std::string message) {
+  return Status(StatusType::PRECONDITION_ERROR, message);
+}
+
+bool Status::ok() const {
+  return type_ == StatusType::OK;
+}
+
+StatusType Status::type() const {
+  return type_;
+}
+
+const std::string& Status::reason() const {
+  return reason_;
+}
+
+void TensorShape::AddDim(int64_t dim) {
+  shape_.push_back(dim);
+}
+
+void TensorShape::AppendShape(TensorShape& other) {
+  for (auto dim : other.shape_) {
+    shape_.push_back(dim);
+  }
+}
+
+const std::string TensorShape::DebugString() const {
+  std::stringstream args;
+  args << "[";
+  for (auto it = shape_.begin(); it != shape_.end(); it++) {
+    if (it != shape_.begin()) {
+      args << ", ";
+    }
+    args << *it;
+  }
+  args << "]";
+  return args.str();
+}
+
+int TensorShape::dims() const {
+  return (int)shape_.size();
+}
+
+int64_t TensorShape::dim_size(int idx) const {
+  assert(idx >= 0);
+  assert(idx < shape_.size());
+  return shape_[idx];
+}
+
+int64_t TensorShape::num_elements() const {
+  int64_t result = 1;
+  for (auto dim : shape_) {
+    result *= dim;
+  }
+  return result;
+}
+
+} // namespace common
+} // namespace horovod
