+++ horovod/common/common.h
@@ -1,0 +1,108 @@
+// Copyright 2018 Uber Technologies, Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// =============================================================================
+
+#ifndef HOROVOD_COMMON_H
+#define HOROVOD_COMMON_H
+
+#include <memory>
+#include <string>
+
+#include "mpi_message.h"
+
+namespace horovod {
+namespace common {
+
+// List of supported frameworks.
+enum Framework { TENSORFLOW };
+
+enum StatusType { OK, UNKNOWN_ERROR, PRECONDITION_ERROR };
+
+class Status {
+public:
+  Status();
+  static Status OK();
+  static Status UnknownError(std::string message);
+  static Status PreconditionError(std::string message);
+  bool ok() const;
+  StatusType type() const;
+  const std::string& reason() const;
+
+private:
+  StatusType type_ = StatusType::OK;
+  std::string reason_ = "";
+  Status(StatusType type, std::string reason);
+};
+
+class TensorShape {
+public:
+  void AddDim(int64_t dim);
+  void AppendShape(TensorShape& other);
+
+  const std::string DebugString() const;
+  int dims() const;
+  int64_t dim_size(int idx) const;
+  int64_t num_elements() const;
+
+  inline bool operator==(const TensorShape& rhs) const {
+    return shape_ == rhs.shape_;
+  }
+
+  inline bool operator!=(const TensorShape& rhs) const {
+    return shape_ != rhs.shape_;
+  }
+
+private:
+  std::vector<int64_t> shape_;
+};
+
+class ReadyEvent {
+public:
+  virtual bool Ready() const = 0;
+  virtual ~ReadyEvent(){};
+};
+
+class OpContext;
+
+class PersistentBuffer {
+public:
+  virtual const char* AccessData(std::shared_ptr<OpContext> context) const = 0;
+  virtual ~PersistentBuffer(){};
+};
+
+class Tensor {
+public:
+  virtual const MPIDataType dtype() const = 0;
+  virtual const TensorShape shape() const = 0;
+  virtual const char* data() const = 0;
+  virtual int64_t size() const = 0;
+  virtual ~Tensor(){};
+};
+
+class OpContext {
+public:
+  // These allocators are fully synchronous, unlike TensorFlow counterparts.
+  virtual Status
+  AllocatePersistent(int64_t size,
+                     std::shared_ptr<PersistentBuffer>* tensor) = 0;
+  virtual Status AllocateOutput(TensorShape shape,
+                                std::shared_ptr<Tensor>* tensor) = 0;
+  virtual Framework framework() const = 0;
+  virtual ~OpContext(){};
+};
+
+} // namespace common
+} // namespace horovod
+
+#endif // HOROVOD_COMMON_H
