+++ horovod/common/operations.h
@@ -1,0 +1,82 @@
+// Copyright 2018 Uber Technologies, Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// =============================================================================
+
+#ifndef HOROVOD_OPERATIONS_H
+#define HOROVOD_OPERATIONS_H
+
+#include <functional>
+
+#include "common.h"
+
+namespace horovod {
+namespace common {
+
+// Device ID used for CPU.
+#define CPU_DEVICE_ID -1
+
+// A callback to call after the MPI communication completes. Since the
+// allreduce and allgather ops are asynchronous, this callback is what resumes
+// computation after the reduction is completed.
+typedef std::function<void(const Status&)> StatusCallback;
+
+// Check that Horovod is initialized.
+Status CheckInitialized();
+
+extern "C" {
+
+// C interface to initialize Horovod.
+void horovod_init();
+
+// C interface to get index of current Horovod process.
+// Returns -1 if Horovod is not initialized.
+int horovod_rank();
+
+// C interface to get index of current Horovod process in the node it is on..
+// Returns -1 if Horovod is not initialized.
+int horovod_local_rank();
+
+// C interface to return number of Horovod processes.
+// Returns -1 if Horovod is not initialized.
+int horovod_size();
+
+// C interface to return number of Horovod processes in the node it is on..
+// Returns -1 if Horovod is not initialized.
+int horovod_local_size();
+}
+
+void EnqueueTensorAllreduce(std::shared_ptr<OpContext> context,
+                            std::shared_ptr<Tensor> tensor,
+                            std::shared_ptr<Tensor> output,
+                            std::shared_ptr<ReadyEvent> ready_event,
+                            const std::string name, const int device,
+                            StatusCallback callback);
+
+void EnqueueTensorAllgather(std::shared_ptr<OpContext> context,
+                            std::shared_ptr<Tensor> tensor,
+                            std::shared_ptr<ReadyEvent> ready_event,
+                            const std::string name, const int device,
+                            StatusCallback callback);
+
+void EnqueueTensorBroadcast(std::shared_ptr<OpContext> context,
+                            std::shared_ptr<Tensor> tensor,
+                            std::shared_ptr<Tensor> output, int root_rank,
+                            std::shared_ptr<ReadyEvent> ready_event,
+                            const std::string name, const int device,
+                            StatusCallback callback);
+
+} // namespace common
+} // namespace horovod
+
+#endif // HOROVOD_OPERATIONS_H
