+++ horovod/common/timeline.h
@@ -1,4 +1,4 @@
-// Copyright 2017 Uber Technologies, Inc. All Rights Reserved.
+// Copyright 2018 Uber Technologies, Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,26 +20,18 @@
 #include <fstream>
 #include <iostream>
 #include <unordered_map>
+#include <mutex>
 
-#include "tensorflow/core/framework/op.h"
-#include "tensorflow/core/framework/op_kernel.h"
-
+#include "common.h"
 #include "mpi_message.h"
 
-using namespace tensorflow;
-
 namespace horovod {
-namespace tensorflow {
+namespace common {
 
 // How frequently Horovod Timeline should be flushed to disk.
 #define TIMELINE_FLUSH_TIME std::chrono::seconds(1)
 
-enum TimelineState {
-  UNKNOWN,
-  NEGOTIATING,
-  TOP_LEVEL,
-  ACTIVITY
-};
+enum TimelineState { UNKNOWN, NEGOTIATING, TOP_LEVEL, ACTIVITY };
 
 // Writes timeline in Chrome Tracing format. Timeline spec is from:
 // https://github.com/catapult-project/catapult/tree/master/tracing
@@ -53,9 +45,10 @@ public:
   void NegotiateEnd(const std::string& tensor_name);
   void Start(const std::string& tensor_name,
              const MPIResponse::ResponseType response_type);
-  void ActivityStart(const std::string& tensor_name, const std::string& activity);
+  void ActivityStart(const std::string& tensor_name,
+                     const std::string& activity);
   void ActivityEnd(const std::string& tensor_name);
-  void End(const std::string& tensor_name, const Tensor* output_tensor);
+  void End(const std::string& tensor_name, const std::shared_ptr<Tensor> tensor);
 
 private:
   void WriteEvent(const std::string& tensor_name, const char phase,
@@ -86,7 +79,7 @@ private:
   std::unordered_map<std::string, TimelineState> tensor_states_;
 };
 
-} // namespace tensorflow
+} // namespace common
 } // namespace horovod
 
 #endif // HOROVOD_TIMELINE_H
