+++ horovod/common/wire/mpi_message_generated.h
@@ -16,13 +16,13 @@
 // automatically generated by the FlatBuffers compiler, do not modify
 
 
-#ifndef FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_TENSORFLOW_WIRE_H_
-#define FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_TENSORFLOW_WIRE_H_
+#ifndef FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_COMMON_WIRE_H_
+#define FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_COMMON_WIRE_H_
 
 #include "flatbuffers/flatbuffers.h"
 
 namespace horovod {
-namespace tensorflow {
+namespace common {
 namespace wire {
 
 struct MPIRequest;
@@ -30,30 +30,30 @@ struct MPIRequest;
 struct MPIResponse;
 
 enum MPIDataType {
-  MPIDataType_TF_MPI_UINT8 = 0,
-  MPIDataType_TF_MPI_INT8 = 1,
-  MPIDataType_TF_MPI_UINT16 = 2,
-  MPIDataType_TF_MPI_INT16 = 3,
-  MPIDataType_TF_MPI_INT32 = 4,
-  MPIDataType_TF_MPI_INT64 = 5,
-  MPIDataType_TF_MPI_FLOAT32 = 6,
-  MPIDataType_TF_MPI_FLOAT64 = 7,
-  MPIDataType_TF_MPI_BOOL = 8,
-  MPIDataType_MIN = MPIDataType_TF_MPI_UINT8,
-  MPIDataType_MAX = MPIDataType_TF_MPI_BOOL
+  MPIDataType_HOROVOD_UINT8 = 0,
+  MPIDataType_HOROVOD_INT8 = 1,
+  MPIDataType_HOROVOD_UINT16 = 2,
+  MPIDataType_HOROVOD_INT16 = 3,
+  MPIDataType_HOROVOD_INT32 = 4,
+  MPIDataType_HOROVOD_INT64 = 5,
+  MPIDataType_HOROVOD_FLOAT32 = 6,
+  MPIDataType_HOROVOD_FLOAT64 = 7,
+  MPIDataType_HOROVOD_BOOL = 8,
+  MPIDataType_MIN = MPIDataType_HOROVOD_UINT8,
+  MPIDataType_MAX = MPIDataType_HOROVOD_BOOL
 };
 
 inline const char **EnumNamesMPIDataType() {
   static const char *names[] = {
-    "TF_MPI_UINT8",
-    "TF_MPI_INT8",
-    "TF_MPI_UINT16",
-    "TF_MPI_INT16",
-    "TF_MPI_INT32",
-    "TF_MPI_INT64",
-    "TF_MPI_FLOAT32",
-    "TF_MPI_FLOAT64",
-    "TF_MPI_BOOL",
+    "HOROVOD_UINT8",
+    "HOROVOD_INT8",
+    "HOROVOD_UINT16",
+    "HOROVOD_INT16",
+    "HOROVOD_INT32",
+    "HOROVOD_INT64",
+    "HOROVOD_FLOAT32",
+    "HOROVOD_FLOAT64",
+    "HOROVOD_BOOL",
     nullptr
   };
   return names;
@@ -202,7 +202,7 @@ inline flatbuffers::Offset<MPIRequest> CreateMPIRequest(
     flatbuffers::FlatBufferBuilder &_fbb,
     int32_t request_rank = 0,
     MPIRequestType request_type = MPIRequestType_ALLREDUCE,
-    MPIDataType tensor_type = MPIDataType_TF_MPI_UINT8,
+    MPIDataType tensor_type = MPIDataType_HOROVOD_UINT8,
     flatbuffers::Offset<flatbuffers::String> tensor_name = 0,
     int32_t root_rank = 0,
     int32_t device = 0,
@@ -222,12 +222,12 @@ inline flatbuffers::Offset<MPIRequest> CreateMPIRequestDirect(
     flatbuffers::FlatBufferBuilder &_fbb,
     int32_t request_rank = 0,
     MPIRequestType request_type = MPIRequestType_ALLREDUCE,
-    MPIDataType tensor_type = MPIDataType_TF_MPI_UINT8,
+    MPIDataType tensor_type = MPIDataType_HOROVOD_UINT8,
     const char *tensor_name = nullptr,
     int32_t root_rank = 0,
     int32_t device = 0,
     const std::vector<int64_t> *tensor_shape = nullptr) {
-  return horovod::tensorflow::wire::CreateMPIRequest(
+  return horovod::common::wire::CreateMPIRequest(
       _fbb,
       request_rank,
       request_type,
@@ -330,7 +330,7 @@ inline flatbuffers::Offset<MPIResponse> CreateMPIResponseDirect(
     const char *error_message = nullptr,
     const std::vector<int32_t> *devices = nullptr,
     const std::vector<int64_t> *tensor_sizes = nullptr) {
-  return horovod::tensorflow::wire::CreateMPIResponse(
+  return horovod::common::wire::CreateMPIResponse(
       _fbb,
       response_type,
       tensor_names ? _fbb.CreateVector<flatbuffers::Offset<flatbuffers::String>>(*tensor_names) : 0,
@@ -340,7 +340,7 @@ inline flatbuffers::Offset<MPIResponse> CreateMPIResponseDirect(
 }
 
 }  // namespace wire
-}  // namespace tensorflow
+}  // namespace common
 }  // namespace horovod
 
-#endif  // FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_TENSORFLOW_WIRE_H_
+#endif  // FLATBUFFERS_GENERATED_MPIMESSAGE_HOROVOD_COMMON_WIRE_H_
