+++ horovod/tensorflow/__init__.py
@@ -31,14 +31,15 @@ from __future__ import print_function
 
 import tensorflow as tf
 
-from horovod.tensorflow.mpi_ops import size
-from horovod.tensorflow.mpi_ops import local_size
-from horovod.tensorflow.mpi_ops import rank
-from horovod.tensorflow.mpi_ops import local_rank
+from horovod.common import init
+from horovod.common import size
+from horovod.common import local_size
+from horovod.common import rank
+from horovod.common import local_rank
+
 from horovod.tensorflow.mpi_ops import allgather
 from horovod.tensorflow.mpi_ops import broadcast
 from horovod.tensorflow.mpi_ops import _allreduce
-from horovod.tensorflow.mpi_ops import init
 
 
 def allreduce(tensor, average=True, device_dense='', device_sparse=''):
