+++ test/test_common.py
@@ -1,0 +1,74 @@
+# Copyright 2016 The TensorFlow Authors. All Rights Reserved.
+# Modifications copyright (C) 2018 Uber Technologies, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# =============================================================================
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+import os
+
+import horovod.common as hvd
+
+
+def mpi_env_rank_and_size():
+    """Get MPI rank and size from environment variables and return them as a
+    tuple of integers.
+
+    Most MPI implementations have an `mpirun` or `mpiexec` command that will
+    run an MPI executable and set up all communication necessary between the
+    different processors. As part of that set up, they will set environment
+    variables that contain the rank and size of the MPI_COMM_WORLD
+    communicator. We can read those environment variables from Python in order
+    to ensure that `hvd.rank()` and `hvd.size()` return the expected values.
+
+    Since MPI is just a standard, not an implementation, implementations
+    typically choose their own environment variable names. This function tries
+    to support several different implementation, but really it only needs to
+    support whatever implementation we want to use for the TensorFlow test
+    suite.
+
+    If this is not running under MPI, then defaults of rank zero and size one
+    are returned. (This is appropriate because when you call MPI_Init in an
+    application not started with mpirun, it will create a new independent
+    communicator with only one process in it.)
+    """
+    rank_env = 'PMI_RANK OMPI_COMM_WORLD_RANK'.split()
+    size_env = 'PMI_SIZE OMPI_COMM_WORLD_SIZE'.split()
+
+    for rank_var, size_var in zip(rank_env, size_env):
+        rank = os.environ.get(rank_var)
+        size = os.environ.get(size_var)
+        if rank is not None and size is not None:
+            return int(rank), int(size)
+
+    # Default to rank zero and size one if there are no environment variables
+    return 0, 1
+
+
+def test_horovod_rank():
+    """Test that the rank returned by hvd.rank() is correct."""
+    true_rank, _ = mpi_env_rank_and_size()
+    hvd.init()
+    rank = hvd.rank()
+    assert true_rank == rank
+
+
+def test_horovod_size():
+    """Test that the size returned by hvd.size() is correct."""
+    _, true_size = mpi_env_rank_and_size()
+    hvd.init()
+    size = hvd.size()
+    assert true_size == size
